% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amoter-package.r
\docType{package}
\name{amoter}
\alias{amoter}
\alias{amoter-package}
\title{amoter: analyse animal movement on realistic terrain}
\description{
The earth isn't flat! Yet, the movement of terrestrial animals is usually
only considered in the x,y-plane. For many types of questions and types of
terrain this is fine, but there are also many cases where elevation matters.
This package aims to help making such analyses.
}
\details{
The package has functions to fetch elevation from a DEM at the track,
calculate attributes along the track related to the elevation change,
calculate important movement metrics for 3D tracks and compare
track orientation along terain features (like contours).
}
