% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal.R
\name{dal}
\alias{dal}
\title{distance along a line}
\usage{
dal(co)
}
\arguments{
\item{co}{a numeric matrix with the coordinates of the line in columns
(a 2D line would result in a 2-column matrix).}
}
\value{
A vector with distance along the line from first point in the input matrix. \code{x} and \code{y}.
}
\description{
distance along a line
}
\examples{
co <- cbind( x=1:5, y=c(3,5,6,4,7))
dal(co)
}
\author{
Emiel van Loon, \email{e.e.vanloon@uva.nl}
}
