% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distrange_pl.R
\name{distrange_pl}
\alias{distrange_pl}
\title{min and maximum distance of a point with a 2D line-segment}
\usage{
distrange_pl(co, pt)
}
\arguments{
\item{co}{a numeric matrix with two columns and two rows, defining a line
segment; the x-coordinates are given in the first and the y-coordinate
in the second column}

\item{pt}{a 2-element vector with the x and y coordinate of the point}
}
\value{
the distance between the point and the two endpoints of the
line segment
}
\description{
min and maximum distance of a point with a 2D line-segment
}
\examples{
co <- cbind( x=c(8,14), y=c(15,11) )
pt <- c(7,15)
distrange_pl(co,pt)
}
