% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ed_resamp.R
\name{ed_resamp}
\alias{ed_resamp}
\title{find x,y coordinates at equal distances along a piecewise linear line}
\usage{
ed_resamp(co, step = 1, endpt = TRUE)
}
\arguments{
\item{co}{a numeric 2-column matrix with the x-coordinates in the first
column and the y-coordinates in the second}

\item{step}{the length between the steps on the line}

\item{endpt}{a boolean value which indicates whether the nodes of the
piecewise linear line have to be included}
}
\value{
A 3-column matrix with in the first two columns x and y coordinates
of the equidistant points and in the third columns an integer which specifies
in which line-segment of co each point falls.
}
\description{
find x,y coordinates at equal distances along a piecewise linear line
}
\examples{
co <- cbind(x=c(1,3,7,8,5),y=c(12,7,6,10,14))

coi1 <- ed_resamp(co, step=1, endpt=FALSE)
coi2 <- ed_resamp(co, step=2, endpt=FALSE)
plot(co[,1],co[,2],type='l',col='red')
points(coi1[,1],coi1[,2],pch=3)
points(coi2[,1],coi2[,2],pch=1,col='blue')

coi1 <- ed_resamp(co, step=1, endpt=TRUE)
coi2 <- ed_resamp(co, step=2, endpt=TRUE)
plot(co[,1],co[,2],type='l',col='red')
points(coi1[,1],coi1[,2],pch=3)
points(coi2[,1],coi2[,2],pch=1,col='blue')
}
