% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idal.R
\name{idal}
\alias{idal}
\title{inverse distance along a line --> still unfinished}
\usage{
idal(co, d)
}
\arguments{
\item{co}{a numeric matrix with the coordinates of the line in columns
(a 2D line would result in a 2-column matrix).}

\item{d}{the distances along the line for which the coordinates are desired}
}
\value{
A vector with distance along the line from first point in the input matrix. \code{x} and \code{y}.
}
\description{
inverse distance along a line --> still unfinished
}
\examples{
co <- cbind( x=1:5, y=c(3,5,6,4,7))
dal(co)
}
