% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intpoint_pl.R
\name{intpoint_pl}
\alias{intpoint_pl}
\title{intersection point with a 2D line, given a reference point and distance}
\usage{
intpoint_pl(co, pt, dist = 1)
}
\arguments{
\item{co}{a numeric 2-column matrix with the two coordinates defining the
line (x-coordinate in the first column and the y-coordinate in the second)}

\item{pt}{a 2-element vector with the x and y coordinate of the point}

\item{dist}{the length between the point and interesection with the line}
}
\value{
A 2-element vector with the x and y coordinate of the intersection point
}
\description{
intersection point with a 2D line, given a reference point and distance
}
\examples{
co <- cbind( x=c(8,14), y=c(15,11) )
pt <- c(7,15)
ip1 <- intpoint_pl(co,pt,dist=2)
ip2 <- intpoint_pl(co,pt,dist=4)
plot(co[,1], co[,2], type='l', xlim=c(5,15), ylim=c(10,17))
points(pt[1], pt[2], col='red')
points(ip1[1],ip1[2], col='blue', pty=3)
points(ip2[1],ip2[2], col='green', pty=3)

# check distance from pt to ip1
sqrt((ip1-pt)\%*\%(ip1-pt))
# check distance from pt to ip2
sqrt((ip2-pt)\%*\%(ip2-pt))
}
