% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prl2d.R
\name{prl2d}
\alias{prl2d}
\title{projection of 2d line on a straight line}
\usage{
prl2d(co, alpha = 0)
}
\arguments{
\item{co}{a numeric two-column matrix with the coordinates of the line in columns}

\item{alpha}{angle of the projection line with respect to the x-axis}
}
\value{
A 2-column matrix with the projected x and y coordinates
('xp' and 'yp').
}
\description{
The orthogonal projection of a line v onto the line s
(both represented by 2d vectors)
is ( (v's)/(s's) ) * s
Here v's is the inner product of vectors v' and s, and s's is the inner
product of s.
The line s is assumed to cross v at the first coordinate in v.
}
\examples{
co <- cbind( x=1:5, y=c(3,5,6,4,7))
prl2d(co)
prl2d(co,alpha=10)
}
